+++ tests/testthat/test_main_evalcurves.R
@@ -3,7 +3,7 @@ context("Main: Curve evaluation")
 #
 
 test_that("run_evalcurve", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
 
   res1 <- run_evalcurve(testset, toolset)
@@ -18,7 +18,7 @@ test_that("run_evalcurve", {
 test_that("run_evalcurve: testset", {
   testset1 <- create_testset("curve", c("c1", "c2"))
   testset2 <- create_testset("bench", c("b10", "i10"))
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
 
   expect_silent(run_evalcurve(testset1, toolset))
 
@@ -31,8 +31,8 @@ test_that("run_evalcurve: testset", {
 
 test_that("run_evalcurve: toolset", {
   testset <- create_testset("curve", c("c1", "c2"))
-  toolset1 <- create_toolset(set_names = "crv5")
-  toolset2 <- create_toolset(set_names = "auc5")
+  toolset1 <- create_toolset(set_names = "crv7")
+  toolset2 <- create_toolset(set_names = "auc7")
   toolset3 <- create_toolset(set_names = "crv6")
   toolset4 <- create_toolset(set_names = "auc6")
 
@@ -61,7 +61,7 @@ test_that("run_evalcurve: toolset & testset", {
 })
 
 test_that("run_evalcurve testscores", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
   res1 <- run_evalcurve(testset, toolset)
 
@@ -71,14 +71,14 @@ test_that("run_evalcurve testscores", {
     "total"
   ))
   expect_true(all(res1$testscores$testset == "c1"))
-  expect_true(all(res1$testscores$toolset == "crv5"))
+  expect_true(all(res1$testscores$toolset == "crv7"))
   expect_true(any(res1$testscores$toolname == "ROCR"))
   expect_true(any(res1$testscores$toolname == "AUCCalculator"))
   expect_true(any(res1$testscores$toolname == "PerfMeas"))
   expect_true(any(res1$testscores$toolname == "PRROC"))
   expect_true(any(res1$testscores$toolname == "precrec"))
 
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", c("c1", "c2"))
   res2 <- run_evalcurve(testset, toolset)
 
@@ -89,7 +89,7 @@ test_that("run_evalcurve testscores", {
   ))
   expect_true(any(res2$testscores$testset == "c1"))
   expect_true(any(res2$testscores$testset == "c2"))
-  expect_true(all(res2$testscores$toolset == "crv5"))
+  expect_true(all(res2$testscores$toolset == "crv7"))
   expect_true(any(res2$testscores$toolname == "ROCR"))
   expect_true(any(res2$testscores$toolname == "AUCCalculator"))
   expect_true(any(res2$testscores$toolname == "PerfMeas"))
@@ -117,7 +117,7 @@ test_that("run_evalcurve testscores", {
 })
 
 test_that("run_evalcurve testsum", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
   res1 <- run_evalcurve(testset, toolset)
 
@@ -128,14 +128,14 @@ test_that("run_evalcurve testsum", {
     "lbl_pos_y2"
   ))
   expect_true(all(res1$testsum$testset == "c1"))
-  expect_true(all(res1$testsum$toolset == "crv5"))
+  expect_true(all(res1$testsum$toolset == "crv7"))
   expect_true(any(res1$testsum$toolname == "ROCR"))
   expect_true(any(res1$testsum$toolname == "AUCCalculator"))
   expect_true(any(res1$testsum$toolname == "PerfMeas"))
   expect_true(any(res1$testsum$toolname == "PRROC"))
   expect_true(any(res1$testsum$toolname == "precrec"))
 
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", c("c1", "c2"))
   res2 <- run_evalcurve(testset, toolset)
 
@@ -147,7 +147,7 @@ test_that("run_evalcurve testsum", {
   ))
   expect_true(any(res2$testsum$testset == "c1"))
   expect_true(any(res2$testsum$testset == "c2"))
-  expect_true(all(res2$testsum$toolset == "crv5"))
+  expect_true(all(res2$testsum$toolset == "crv7"))
   expect_true(any(res2$testsum$toolname == "ROCR"))
   expect_true(any(res2$testsum$toolname == "AUCCalculator"))
   expect_true(any(res2$testsum$toolname == "PerfMeas"))
@@ -176,7 +176,7 @@ test_that("run_evalcurve testsum", {
 })
 
 test_that("run_evalcurve catres", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
   res1 <- run_evalcurve(testset, toolset)
 
@@ -185,7 +185,7 @@ test_that("run_evalcurve catres", {
     "toolname", "success", "total", "label"
   ))
   expect_true(all(res1$catres$testset == "c1"))
-  expect_true(all(res1$catres$toolset == "crv5"))
+  expect_true(all(res1$catres$toolset == "crv7"))
   expect_true(any(res1$catres$testcat == "SE"))
   expect_true(any(res1$catres$testcat == "Ip"))
   expect_true(any(res1$catres$testcat == "Rg"))
@@ -195,7 +195,7 @@ test_that("run_evalcurve catres", {
   expect_true(any(res1$catres$toolname == "PRROC"))
   expect_true(any(res1$catres$toolname == "precrec"))
 
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", c("c1", "c2"))
   res2 <- run_evalcurve(testset, toolset)
 
@@ -205,7 +205,7 @@ test_that("run_evalcurve catres", {
   ))
   expect_true(any(res2$catres$testset == "c1"))
   expect_true(any(res2$catres$testset == "c2"))
-  expect_true(all(res2$catres$toolset == "crv5"))
+  expect_true(all(res2$catres$toolset == "crv7"))
   expect_true(any(res2$catres$testcat == "SE"))
   expect_true(any(res2$catres$testcat == "Ip"))
   expect_true(any(res2$catres$testcat == "Rg"))
@@ -238,7 +238,7 @@ test_that("run_evalcurve catres", {
 })
 
 test_that("run_evalcurve basepoints", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
   res1 <- run_evalcurve(testset, toolset)
 
@@ -247,7 +247,7 @@ test_that("run_evalcurve basepoints", {
   expect_true(all(res1$basepoints$y >= 0) && all(res1$basepoints$y <= 1))
   expect_true(all(res1$basepoints$testset == "c1"))
 
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", c("c1", "c2"))
   res2 <- run_evalcurve(testset, toolset)
 
@@ -269,7 +269,7 @@ test_that("run_evalcurve basepoints", {
 })
 
 test_that("run_evalcurve predictions", {
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", "c1")
   res1 <- run_evalcurve(testset, toolset)
 
@@ -278,7 +278,7 @@ test_that("run_evalcurve predictions", {
     "x", "y"
   ))
   expect_true(all(res1$predictions$testset == "c1"))
-  expect_true(all(res1$predictions$toolset == "crv5"))
+  expect_true(all(res1$predictions$toolset == "crv7"))
   expect_true(any(res1$predictions$toolname == "ROCR"))
   if (length(unique(res1$predictions$toolset)) == 5) {
     expect_true(any(res1$predictions$toolname == "AUCCalculator"))
@@ -287,7 +287,7 @@ test_that("run_evalcurve predictions", {
   expect_true(any(res1$predictions$toolname == "PRROC"))
   expect_true(any(res1$predictions$toolname == "precrec"))
 
-  toolset <- create_toolset(set_names = "crv5")
+  toolset <- create_toolset(set_names = "crv7")
   testset <- create_testset("curve", c("c1", "c2"))
   res2 <- run_evalcurve(testset, toolset)
 
@@ -297,7 +297,7 @@ test_that("run_evalcurve predictions", {
   ))
   expect_true(any(res2$predictions$testset == "c1"))
   expect_true(any(res2$predictions$testset == "c2"))
-  expect_true(all(res2$predictions$toolset == "crv5"))
+  expect_true(all(res2$predictions$toolset == "crv7"))
   expect_true(any(res2$predictions$toolname == "ROCR"))
   if (length(unique(res1$predictions$toolset)) == 5) {
     expect_true(any(res2$predictions$toolname == "AUCCalculator"))